code Refactoring

+++ src/main/java/com/haufe/beercatalogue/service/ManufacturerService.java
@@ -6,17 +6,29 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.haufe.beercatalogue.domain.Manufacturer;
+import com.haufe.beercatalogue.exception.ConflictException;
 import com.haufe.beercatalogue.exception.NotFoundException;
+import com.haufe.beercatalogue.repository.AppUserRepository;
+import com.haufe.beercatalogue.repository.BeerRepository;
 import com.haufe.beercatalogue.repository.ManufacturerRepository;
 
 @Service
 @Transactional
 public class ManufacturerService {
     private final ManufacturerRepository manufacturerRepository;
+    private final BeerRepository beerRepository;
+    private final AppUserRepository appUserRepository;
     private final AccessService accessService;
 
-    public ManufacturerService(final ManufacturerRepository manufacturerRepository, final AccessService accessService) {
+    public ManufacturerService(
+            final ManufacturerRepository manufacturerRepository,
+            final BeerRepository beerRepository,
+            final AppUserRepository appUserRepository,
+            final AccessService accessService
+    ) {
         this.manufacturerRepository = manufacturerRepository;
+        this.beerRepository = beerRepository;
+        this.appUserRepository = appUserRepository;
         this.accessService = accessService;
     }
 
@@ -47,6 +59,12 @@ public class ManufacturerService {
     public void delete(final Long id) {
         accessService.requireAdminOrOwnManufacturer(id);
         final var manufacturer = findById(id);
+        if (beerRepository.existsByManufacturer_Id(id)) {
+            throw new ConflictException("Manufacturer with id " + id + " cannot be deleted because beers are linked to it");
+        }
+        if (appUserRepository.existsByManufacturer_Id(id)) {
+            throw new ConflictException("Manufacturer with id " + id + " cannot be deleted because users are linked to it");
+        }
         manufacturerRepository.delete(manufacturer);
     }
 }
